refactor: Extract breakpoint helpers from withBreakpoints
